+++ backend/canisters/notifications/api/src/lifecycle/init.rs
@@ -6,7 +6,7 @@ use types::{BuildVersion, CanisterId};
 pub struct Args {
     pub notifications_index_canister_id: CanisterId,
     pub push_service_principals: Vec<Principal>,
-    pub authorizers: Vec<CanisterId>,
+    pub local_user_index_canister_id: Principal,
     pub cycles_dispenser_canister_id: CanisterId,
     pub wasm_version: BuildVersion,
     pub test_mode: bool,